Add CI workflow and pyright config; fix type findings in the test suite
CI (.github/workflows/tests.yml)
-------------------------------
Two jobs on push / pull_request / manual dispatch:

  * test -- runs the suite with VRP_FULL_MATRIX=1, i.e. the exhaustive operand sweeps rather than
    the reduced set used for local speed (~20s). Blocking. Pinned to Python 3.14, which is a hard
    requirement rather than a preference: the model uses PEP 695 generics (class Foo[T],
    type X = ...) needing 3.12+, and depends on PEP 649 deferred annotation evaluation for the
    forward references throughout the core model, which needs 3.14.
  * typecheck -- runs pyright, non-blocking. The codebase still carries known findings that are
    tracked rather than fixed, so failing the build on them would only train people to ignore CI.
    The value is the diff: a NEW finding in a pull request is the thing worth looking at.

pyrightconfig.json
------------------
Scopes checking to the solver modules plus tests, and pins pythonVersion 3.14 -- the single most
important setting here, since without it pyright assumes eager annotation evaluation and reports
~76 spurious forward-reference errors that PEP 649 makes impossible. That alone took the raw
count from 117 to 51. reportIncompatibleVariableOverride is disabled because the sentinel classes
deliberately narrow their base's annotations (FirstRoute.prev_route: None over
VehicleNode.prev_route: VehicleNode | None), and reportIncompatibleMethodOverride is a warning
rather than an error for the remaining non-operator cases.

Fixes for findings this surfaced in code added by the test suite:
  * _harness.term_deltas built its result by splatting a generator, which widens every field to
    the union of all field types; written out field by field so each keeps its declared type.
  * The operator contract helper was a bare mixin calling self.assertEqual, which is not declared
    anywhere on it. It is now OperatorContractBase(SeededTestCase) -- it defines no test_* methods
    so discovery still collects nothing from it directly.
  * SimAnnVRPSolver.max_time was annotated int but holds a duration in seconds compared against
    float elapsed time, and the suite passes sub-second budgets. Now float.

Suite: 18 tests, green locally under both the default and exhaustive matrix settings.
Remaining pyright state: 9 errors, 11 warnings, none in tests/.

+++ SimAnn_VRP_Solver.py
@@ -49,7 +49,7 @@ def argmin(values):
 #    can see, since it only manifests on the apply->revert path.
 
 class SimAnnVRPSolver:
-    def __init__(self, sln: FullSolution, max_time: int = 120):
+    def __init__(self, sln: FullSolution, max_time: float = 120):
         self.sln = sln
         self.operators: list[Operator] = []
